requirements.txt
asgiref==3.8.1
cachetools==5.3.3
chardet==5.2.0
colorama==0.4.6
coverage==7.5.4
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.8
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-isort==6.1.1
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
PyJWT==2.8.0
pyproject-api==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.0.2
snowballstemmer==2.2.0
sqlparse==0.5.0
tox==4.11.4
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
